game: graph structs are builder internals, not header API
